+++ packages/cli/cli/src/commands/exp/generate/index.ts
@@ -1,0 +1,1 @@
+export * from "./generateAPIWorkspacesLocal";
