+++ packages/cli/cli/src/commands/exp/generate/types.ts
@@ -1,0 +1,11 @@
+import { IntermediateRepresentation } from "@fern-api/ir-sdk";
+import { getGeneratorConfig } from "@fern-api/local-workspace-runner";
+
+export type SdkGeneratorCLI = new () => {
+    runProgrammatically: (
+        ir: IntermediateRepresentation,
+        generatorConfig: ReturnType<typeof getGeneratorConfig>
+    ) => Promise<void>;
+};
+
+export type SdkGeneratorCLILoader = (version: string) => Promise<SdkGeneratorCLI>;
